Refactor cmdstore CLI structure and remove unused code
- Consolidated imports by moving main functionality to dedicated modules for better organization.
- Removed redundant code related to ANSI styling and command handling, streamlining the CommandStore class.
- Enhanced clarity and maintainability of the codebase by reducing complexity in the main script.
- Updated the __all__ variable to reflect the new module structure.
